Configurations, exceptions, constraints, services

+++ src/main/java/com/sling/technology/infraestructure/adapter/in/web/CountResponseDTO.java
@@ -3,6 +3,5 @@ package com.sling.technology.infraestructure.adapter.in.web;
 public record CountResponseDTO(
         String searchId,
         SearchRequestDTO search,
-        Integer count
-) {
+        Long count) {
 }
